refactor the change request DAOs to remove the interfaces
OCD-3122
